Move API key endpoints to controller
