+++ index.js
@@ -531,98 +531,6 @@ fastify.delete('/api/files/:fileId/versions/:versionNumber', async (req, reply) 
 
 
 
-// --- API KEY ENDPOINTS ---
-fastify.post('/api/v1/keys', async (req, reply) => {
-  try {
-    // Verify JWT token
-    const token = req.headers.authorization?.split(' ')[1];
-    if (!token) {
-      return reply.code(401).send({ error: 'Authentication required' });
-    }
-
-    const decoded = jwt.verify(token, JWT_SECRET);
-    const { description } = req.body;
-
-    if (!description) {
-      return reply.code(400).send({ error: 'Description is required' });
-    }
-
-    const key = await generateApiKey(decoded.username, description);
-    
-    // Log the activity
-    await logActivity(req, 'generate-api-key', { description });
-
-    return reply.send({ key });
-  } catch (err) {
-    if (err.name === 'JsonWebTokenError') {
-      return reply.code(401).send({ error: 'Invalid token' });
-    }
-    throw err;
-  }
-});
-
-fastify.get('/api/v1/keys', async (req, reply) => {
-  try {
-    // Verify JWT token
-    const token = req.headers.authorization?.split(' ')[1];
-    if (!token) {
-      return reply.code(401).send({ error: 'Authentication required' });
-    }
-
-    const decoded = jwt.verify(token, JWT_SECRET);
-    
-    // Get all active API keys for the user
-    const keys = await apiKeysColl.find(
-      { username: decoded.username, isActive: true },
-      { projection: { key: 0 } } // Don't send the actual keys
-    ).toArray();
-
-    return reply.send({ keys });
-  } catch (err) {
-    if (err.name === 'JsonWebTokenError') {
-      return reply.code(401).send({ error: 'Invalid token' });
-    }
-    throw err;
-  }
-});
-
-fastify.delete('/api/v1/keys/:keyId', async (req, reply) => {
-  try {
-    // Verify JWT token
-    const token = req.headers.authorization?.split(' ')[1];
-    if (!token) {
-      return reply.code(401).send({ error: 'Authentication required' });
-    }
-
-    const decoded = jwt.verify(token, JWT_SECRET);
-    const { keyId } = req.params;
-
-    // Verify the key belongs to the user and deactivate it
-    const result = await apiKeysColl.updateOne(
-      { 
-        _id: new ObjectId(keyId), 
-        username: decoded.username,
-        isActive: true 
-      },
-      { $set: { isActive: false } }
-    );
-
-    if (result.matchedCount === 0) {
-      return reply.code(404).send({ error: 'API key not found' });
-    }
-
-    // Log the activity
-    await logActivity(req, 'revoke-api-key', { keyId });
-
-    return reply.send({ success: true });
-  } catch (err) {
-    if (err.name === 'JsonWebTokenError') {
-      return reply.code(401).send({ error: 'Invalid token' });
-    }
-    throw err;
-  }
-});
-
 // --- CONFIGURATION ---
 const PORT             = process.env.PORT || 3000;
 const JWT_SECRET       = process.env.JWT_SECRET;
@@ -1021,6 +929,22 @@ function consumeDownloadToken(token) {
   downloadTokens.delete(token);
   return data;
 }
+
+// Register API v1 routes
+fastify.register(require('./controllers/apiV1'), {
+  validateApiKey,
+  apiKeysColl,
+  foldersColl,
+  s3,
+  jwt,
+  JWT_SECRET,
+  generateApiKey,
+  logActivity,
+  ObjectId,
+  COMPRESSION_ENABLED,
+  COMPRESSION_THRESHOLD,
+  COMPRESSION_LEVEL
+});
 
 
 
@@ -1787,394 +1711,6 @@ fastify.get('/api/view-file/:folderId/*', async (req, reply) => {
   }
 });
 
-// GET /api/v1/file/:fileId
-fastify.get('/api/v1/file/:fileId', async (req, reply) => {
-  try {
-    const { fileId } = req.params;
-    const apiKey = req.headers['x-api-key'];
-    const range = req.headers.range;
-
-    // Parse file ID with proper URL decoding
-    let folderId, filename;
-    try {
-      // First try the standard format (folderId:filename)
-      const parts = decodeURIComponent(fileId).split(':', 2);
-      if (parts.length === 2) {
-        [folderId, filename] = parts;
-      } else {
-        // If no colon found, try to extract from the last occurrence of a folder ID pattern
-        // This handles cases where the filename might contain colons
-        const folderIdPattern = /^[a-f0-9]{32}$/; // Assuming folder IDs are 32-character hex
-        const lastFolderIdIndex = fileId.lastIndexOf('/');
-        if (lastFolderIdIndex !== -1) {
-          const potentialFolderId = fileId.slice(lastFolderIdIndex + 1, lastFolderIdIndex + 33);
-          if (folderIdPattern.test(potentialFolderId)) {
-            folderId = potentialFolderId;
-            filename = fileId.slice(lastFolderIdIndex + 34);
-          }
-        }
-      }
-    } catch (err) {
-      fastify.log.warn('Error parsing fileId:', { fileId, error: err.message });
-      return reply.code(400).send({ error: 'Invalid file ID format' });
-    }
-
-    if (!folderId || !filename) {
-      return reply.code(400).send({ error: 'Invalid file ID format' });
-    }
-
-    // Load folder metadata from MongoDB
-    const meta = await foldersColl.findOne({ folderId });
-    if (!meta) {
-      return reply.code(404).send({ error: 'Folder not found' });
-    }
-
-    // Check permissions for private folders
-    if (!meta.isPublic) {
-      const keyData = await validateApiKey(apiKey);
-      if (!keyData) {
-        return reply.code(401).send({ error: 'Invalid API key' });
-      }
-
-      // Check if user is owner or has permission
-      const hasAccess = meta.owner === keyData.username || 
-                       (meta.permissions && 
-                        meta.permissions[keyData.username] && 
-                        meta.permissions[keyData.username].download);
-
-      if (!hasAccess) {
-        return reply.code(403).send({ error: 'Access denied' });
-      }
-
-      // Update usage count
-      await apiKeysColl.updateOne(
-        { _id: keyData._id },
-        { $inc: { usageCount: 1 } }
-      );
-    }
-
-    // Get file metadata from S3
-    const key = `folders/${folderId}/${filename}`;
-    let headRes;
-    try {
-      headRes = await s3.send(new HeadObjectCommand({
-        Bucket: process.env.S3_BUCKET_NAME,
-        Key: key
-      }));
-    } catch (err) {
-      if (err.name === 'NotFound' || err.name === 'NoSuchKey') {
-        // Try alternative encoding if file not found
-        try {
-          const alternativeKey = `folders/${folderId}/${encodeURIComponent(filename)}`;
-          headRes = await s3.send(new HeadObjectCommand({
-            Bucket: process.env.S3_BUCKET_NAME,
-            Key: alternativeKey
-          }));
-        } catch (altErr) {
-          fastify.log.error('File not found with both original and encoded filename:', {
-            folderId,
-            filename,
-            error: err.message,
-            altError: altErr.message
-          });
-          return reply.code(404).send({ error: 'File not found' });
-        }
-      } else {
-        throw err;
-      }
-    }
-
-    // Check if file is compressed
-    const isCompressed = headRes.Metadata?.['is-compressed'] === 'true';
-    const originalContentType = headRes.Metadata?.['original-content-type'];
-
-    // Determine MIME type and content disposition
-    const mimeType = isCompressed ? originalContentType : (mime.lookup(filename) || 'application/octet-stream');
-    const totalSize = headRes.ContentLength;
-    const fileExtension = path.extname(filename).toLowerCase();
-    
-    // Determine if file should be displayed inline or downloaded
-    const inlineTypes = ['.pdf', '.jpg', '.jpeg', '.png', '.gif', '.svg', '.webp', '.mp4', '.webm', '.mp3', '.wav'];
-    const shouldDisplayInline = inlineTypes.includes(fileExtension);
-    const disposition = shouldDisplayInline ? 'inline' : 'attachment';
-
-    // Handle range requests for partial content
-    let rangeStart = 0;
-    let rangeEnd = totalSize - 1;
-    let isRangeRequest = false;
-
-    if (range && totalSize > 0) {
-      const rangeMatch = /bytes=(\d+)-(\d*)/.exec(range);
-      if (rangeMatch) {
-        rangeStart = parseInt(rangeMatch[1], 10);
-        if (rangeMatch[2]) {
-          rangeEnd = Math.min(parseInt(rangeMatch[2], 10), totalSize - 1);
-        }
-        
-        // Validate range
-        if (rangeStart >= totalSize || rangeStart > rangeEnd) {
-          return reply
-            .code(416)
-            .header('Content-Range', `bytes */${totalSize}`)
-            .send({ error: 'Range not satisfiable' });
-        }
-        
-        isRangeRequest = true;
-      }
-    }
-
-    // Get file content from S3
-    const getObjectParams = {
-      Bucket: process.env.S3_BUCKET_NAME,
-      Key: key
-    };
-
-    if (isRangeRequest) {
-      getObjectParams.Range = `bytes=${rangeStart}-${rangeEnd}`;
-    }
-
-    const s3Response = await s3.send(new GetObjectCommand(getObjectParams));
-    let stream = s3Response.Body;
-
-    // Handle stream errors
-    stream.on('error', (err) => {
-      fastify.log.error('Stream error while serving file:', {
-        error: err.message,
-        stack: err.stack,
-        fileId: req.params.fileId,
-        key,
-        rangeStart,
-        rangeEnd
-      });
-    });
-
-    // Decompress if needed
-    if (isCompressed) {
-      const gunzip = zlib.createGunzip();
-      stream = stream.pipe(gunzip);
-    }
-
-    // Set response headers
-    const responseHeaders = {
-      'Content-Type': mimeType,
-      'Content-Disposition': `${disposition}; filename="${encodeURIComponent(filename)}"`,
-      'Cache-Control': 'public, max-age=3600',
-      'ETag': headRes.ETag,
-      'Last-Modified': headRes.LastModified
-    };
-
-    if (isRangeRequest) {
-      responseHeaders['Accept-Ranges'] = 'bytes';
-      responseHeaders['Content-Range'] = `bytes ${rangeStart}-${rangeEnd}/${totalSize}`;
-      responseHeaders['Content-Length'] = rangeEnd - rangeStart + 1;
-    } else {
-      responseHeaders['Content-Length'] = totalSize;
-    }
-
-    return reply
-      .code(isRangeRequest ? 206 : 200)
-      .headers(responseHeaders)
-      .send(stream);
-
-  } catch (err) {
-    fastify.log.error('Error serving file:', {
-      error: err.message,
-      stack: err.stack,
-      fileId: req.params.fileId,
-      key: `folders/${folderId}/${filename}`,
-      s3Error: err.name,
-      s3Code: err.$metadata?.httpStatusCode,
-      s3RequestId: err.$metadata?.requestId
-    });
-    
-    if (err.name === 'NoSuchKey' || err.name === 'NotFound') {
-      return reply.code(404).send({ error: 'File not found' });
-    }
-    
-    return reply.code(500).send({ error: 'Failed to serve file' });
-  }
-});
-
-
-// GET /api/v1/folder/:folderId
-fastify.get('/api/v1/folder/:folderId', async (req, reply) => {
-  try {
-    const { folderId } = req.params;
-    const folder = await foldersColl.findOne({ folderId });
-    if (!folder) return reply.notFound('Folder not found');
-    return folder;
-  } catch (err) {
-    fastify.log.error('Error getting folder:', err);
-    return reply.internalServerError('Error retrieving folder');
-  }
-});
-
-// GET /api/v1/latest/:folderId
-fastify.get('/api/v1/latest/:folderId', async (req, reply) => {
-  const { folderId } = req.params;
-  const apiKey = req.headers['x-api-key'];
-  const folder = await foldersColl.findOne({ folderId });
-  if (!folder) return reply.notFound('Folder not found');
-
-  const keyData = await validateApiKey(apiKey);
-  if (!keyData) return reply.code(401).send({ error: 'Invalid API key' });
-
-  if (!folder.isPublic) {
-    const hasAccess = folder.owner === keyData.username ||
-      (folder.permissions?.[keyData.username]?.download);
-    if (!hasAccess) return reply.forbidden('Access denied');
-  }
-
-  const data = await s3.send(new ListObjectsV2Command({
-    Bucket: process.env.S3_BUCKET_NAME,
-    Prefix: `folders/${folderId}/`
-  }));
-  if (!data.Contents?.length) return reply.notFound('No files in folder');
-
-  const latest = data.Contents
-    .filter(obj => obj.Key !== `folders/${folderId}/`)
-    .sort((a,b) => b.LastModified - a.LastModified)[0];
-  const filename = latest.Key.slice(`folders/${folderId}/`.length);
-
-  await apiKeysColl.updateOne({ _id: keyData._id }, { $inc: { usageCount: 1 } });
-
-  return reply.send({
-    id: `${folderId}:${filename}`,
-    name: filename,
-    size: latest.Size,
-    type: mime.lookup(filename) || 'application/octet-stream',
-    lastModified: latest.LastModified.toISOString(),
-    url: `https://hackclub.maksimmalbasa.in.rs/api/v1/file/${folderId}:${filename}`
-  });
-});
-
-
-// POST /api/v1/upload
-fastify.post('/api/v1/upload', async (req, reply) => {
-  const apiKey = req.headers['x-api-key'];
-  const folderId = req.query.folderId;
-  if (!folderId) return reply.badRequest('folderId query parameter is required');
-
-  const keyData = await validateApiKey(apiKey);
-  if (!keyData) return reply.code(401).send({ error: 'Invalid API key' });
-
-  const folder = await foldersColl.findOne({ folderId });
-  if (!folder) return reply.notFound('Folder not found');
-
-  const hasUpload = folder.owner === keyData.username ||
-    (folder.permissions?.[keyData.username]?.upload);
-  if (!hasUpload) return reply.forbidden('No upload permission');
-
-  const uploadedFiles = [];
-  for await (const file of await req.files()) {
-    if (file.file.truncated) return reply.entityTooLarge('File too large');
-    const filename = `${Date.now()}-${file.filename}`;
-    let fileBuffer = await file.toBuffer();
-    let contentType = file.mimetype || mime.lookup(filename) || 'application/octet-stream';
-    let contentLength = fileBuffer.length;
-    let isCompressed = false;
-
-    // Apply compression if enabled and file is large enough
-    if (COMPRESSION_ENABLED && fileBuffer.length >= COMPRESSION_THRESHOLD) {
-      try {
-        const compressed = await new Promise((resolve, reject) => {
-          zlib.gzip(fileBuffer, { level: COMPRESSION_LEVEL }, (err, result) => {
-            if (err) reject(err);
-            else resolve(result);
-          });
-        });
-        
-        // Only use compression if it actually reduced the size
-        if (compressed.length < fileBuffer.length) {
-          fileBuffer = compressed;
-          contentType = 'application/gzip';
-          contentLength = compressed.length;
-          isCompressed = true;
-        }
-      } catch (err) {
-        fastify.log.warn('Compression failed:', err);
-        // Continue with uncompressed file
-      }
-    }
-
-    await s3.send(new PutObjectCommand({
-      Bucket: process.env.S3_BUCKET_NAME,
-      Key: `folders/${folderId}/${filename}`,
-      Body: fileBuffer,
-      ContentType: contentType,
-      ContentLength: contentLength,
-      Metadata: {
-        'original-content-type': file.mimetype || mime.lookup(filename) || 'application/octet-stream',
-        'is-compressed': isCompressed.toString()
-      }
-    }));
-
-    uploadedFiles.push({
-      id: `${folderId}:${filename}`,
-      name: filename,
-      size: contentLength,
-      type: contentType,
-      isCompressed,
-      url: `https://hackclub.maksimmalbasa.in.rs/api/v1/file/${folderId}:${filename}`
-    });
-  }
-
-  await apiKeysColl.updateOne({ _id: keyData._id }, { $inc: { usageCount: 1 } });
-  return reply.send({ files: uploadedFiles });
-});
-
-fastify.delete('/api/v1/delete/:folderId/:filename', async (req, reply) => {
-  const { folderId, filename } = req.params;
-  const apiKey = req.headers['x-api-key'];
-  const keyData = await validateApiKey(apiKey);
-  if (!keyData) return reply.code(401).send({ error: 'Invalid API key' });
-
-  const folder = await foldersColl.findOne({ folderId });
-  if (!folder) return reply.notFound('Folder not found');
-
-  const hasDelete = folder.owner === keyData.username ||
-    (folder.permissions?.[keyData.username]?.delete);
-  if (!hasDelete) return reply.forbidden('Access denied');
-
-  const key = `folders/${folderId}/${filename}`;
-  try {
-    await s3.send(new DeleteObjectCommand({ Bucket: process.env.S3_BUCKET_NAME, Key: key }));
-    return reply.send({ message: 'File deleted' });
-  } catch {
-    return reply.notFound('File not found');
-  }
-});
-// GET /api/v1/usage
-fastify.get('/api/v1/usage', async (req, reply) => {
-  const apiKey = req.headers['x-api-key'];
-  const keyData = await validateApiKey(apiKey);
-  if (!keyData) return reply.code(401).send({ error: 'Invalid API key' });
-
-  const userFolders = await foldersColl.find({ owner: keyData.username }).toArray();
-  let totalSize = 0, totalFiles = 0;
-  for (const folder of userFolders) {
-    const data = await s3.send(new ListObjectsV2Command({
-      Bucket: process.env.S3_BUCKET_NAME,
-      Prefix: `folders/${folder.folderId}/`
-    }));
-    const contents = data.Contents || [];
-    totalFiles += contents.length;
-    totalSize += contents.reduce((sum, obj) => sum + obj.Size, 0);
-  }
-
-  await apiKeysColl.updateOne({ _id: keyData._id }, { $inc: { usageCount: 1 } });
-  return reply.send({
-    apiKey: {
-      created: keyData.created,
-      lastUsed: keyData.lastUsed,
-      totalRequests: keyData.usageCount
-    },
-    storage: {
-      files: totalFiles,
-      totalSize
-    }
-  });
-});
 
 // DELETE /api/delete-file/:folderId/*
 fastify.delete('/api/delete-file/:folderId/*', { preHandler: [fastify.authenticate] }, async (req, reply) => {
